# docs/openapi.yaml
components:
  schemas:
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          examples:
            - Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          examples:
            - https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          examples:
            - 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          examples:
            - Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          examples:
            - https://example.com/errors/example
          format: uri
          type: string
      type: object
    GPU:
      additionalProperties: false
      properties:
        device:
          type: string
        first_seen:
          format: date-time
          type: string
        gpu_id:
          type: string
        hostname:
          type: string
        id:
          type: string
        last_seen:
          format: date-time
          type: string
        model_name:
          type: string
        uuid:
          type: string
      required:
        - id
        - hostname
        - gpu_id
        - device
        - model_name
        - first_seen
        - last_seen
      type: object
    HealthOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/HealthOutputBody.json
          format: uri
          readOnly: true
          type: string
        status:
          type: string
      required:
        - status
      type: object
    ListGPUsOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/ListGPUsOutputBody.json
          format: uri
          readOnly: true
          type: string
        items:
          items:
            $ref: "#/components/schemas/GPU"
          type:
            - array
            - "null"
      required:
        - items
      type: object
    Telemetry:
      additionalProperties: false
      properties:
        assignment_generation:
          format: int64
          minimum: 0
          type: integer
        container:
          type: string
        cycle_id:
          format: int64
          minimum: 0
          type: integer
        dataset_id:
          type: string
        device:
          type: string
        event_id:
          type: string
        gpu_id:
          type: string
        hostname:
          type: string
        ingested_at:
          format: date-time
          type: string
        labels_raw:
          type: string
        metric_name:
          type: string
        model_name:
          type: string
        namespace:
          type: string
        observed_at:
          format: date-time
          type: string
        pod:
          type: string
        queue_offset:
          format: int64
          minimum: 0
          type: integer
        queue_partition:
          format: int32
          minimum: 0
          type: integer
        row_number:
          format: int64
          minimum: 0
          type: integer
        source_event_key:
          type: string
        source_timestamp:
          format: date-time
          type: string
        streamer_id:
          type: string
        uuid:
          type: string
        value:
          format: double
          type: number
      required:
        - event_id
        - source_event_key
        - source_timestamp
        - observed_at
        - metric_name
        - gpu_id
        - device
        - model_name
        - hostname
        - value
        - labels_raw
        - dataset_id
        - cycle_id
        - assignment_generation
        - row_number
        - streamer_id
        - queue_partition
        - queue_offset
      type: object
    TelemetryPage:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/TelemetryPage.json
          format: uri
          readOnly: true
          type: string
        items:
          items:
            $ref: "#/components/schemas/Telemetry"
          type:
            - array
            - "null"
        next_cursor:
          type: string
      required:
        - items
      type: object
info:
  description: Queries all persisted GPU telemetry, including original source timestamps and dynamically assigned observation timestamps.
  title: GPU Telemetry API
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/v1/gpus:
    get:
      operationId: list-gpus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListGPUsOutputBody"
          description: OK
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Service Unavailable
        "504":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Gateway Timeout
      summary: List known GPUs
      tags:
        - Telemetry
  /api/v1/gpus/{id}/telemetry:
    get:
      description: Returns all matched fields from every persisted cycle. Time bounds are inclusive and apply to the dynamically assigned observed_at timestamp. Use event_id for an exact row.
      operationId: get-gpu-telemetry
      parameters:
        - description: Stable GPU key (UUID when available, otherwise hostname/GPU ID)
          in: path
          name: id
          required: true
          schema:
            description: Stable GPU key (UUID when available, otherwise hostname/GPU ID)
            type: string
        - description: Inclusive RFC3339 observation-time lower bound
          explode: false
          in: query
          name: start_time
          schema:
            description: Inclusive RFC3339 observation-time lower bound
            type: string
        - description: Inclusive RFC3339 observation-time upper bound
          explode: false
          in: query
          name: end_time
          schema:
            description: Inclusive RFC3339 observation-time upper bound
            type: string
        - description: Return one exact event by its event ID
          explode: false
          in: query
          name: event_id
          schema:
            description: Return one exact event by its event ID
            type: string
        - description: Filter by metric name
          explode: false
          in: query
          name: metric_name
          schema:
            description: Filter by metric name
            type: string
        - description: Page size; omit to return every match up to the server safety ceiling
          explode: false
          in: query
          name: limit
          schema:
            description: Page size; omit to return every match up to the server safety ceiling
            format: int64
            maximum: 1000
            minimum: 0
            type: integer
        - description: Opaque snapshot cursor returned by an earlier page
          explode: false
          in: query
          name: cursor
          schema:
            description: Opaque snapshot cursor returned by an earlier page
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryPage"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Bad Request
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Not Found
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Unprocessable Entity
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Service Unavailable
        "504":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Gateway Timeout
      summary: Query a GPU's telemetry
      tags:
        - Telemetry
  /healthz:
    get:
      operationId: liveness
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Operations
  /readyz:
    get:
      operationId: readiness
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      tags:
        - Operations